+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+### Added
+- Support for unclosed paths.
+- Private parameter `closed_path` to set path type when receiving it from `path` topic.
+
 ### Changed
 - Do not recompute the path when picking up a point.