+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Support for unclosed paths.
 - Private parameter `closed_path` to set path type when receiving it from `path` topic.
+- Private parameter `input_file` to load initial points from a `.npy` file.
+- Private parameter `delimiter` to load initial points from a `.csv` instead.
 
 ### Changed
 - Do not recompute the path when picking up a point.